+++ modules/vocoders/nsf_hifigan.py
@@ -87,7 +87,7 @@ class NsfHifiGAN(BaseVocoder):
                 y = self.model(c, f0).view(-1)
             else:
                 y = self.model(c).view(-1)
-        wav_out = y.cpu().float().numpy()
+        wav_out = y.cpu().numpy()
         return wav_out
 
     @staticmethod
